Allowing floating point time limits for the algorithm. In addition, calculating elapsed time relatively to a startTime parameter located in Params.h, and initialized when initially constructing the Params object and reading the instance. This permits, if needed, running HGS in a loop on several instances/seed while calculating correctly the time limit.

+++ Program/Genetic.h
@@ -42,7 +42,7 @@ private:
 public:
 
     // Running the genetic algorithm until maxIterNonProd consecutive iterations or a time limit
-    void run(int maxIterNonProd, int timeLimit) ;
+    void run(int maxIterNonProd, double timeLimit) ;
 
 	// Constructor
 	Genetic(Params * params, Split * split, Population * population, LocalSearch * localSearch);
